Endpoint update movement and fix in update product

+++ server/app/types/movement.go
@@ -1,7 +1,7 @@
 package types
 
 type Movement struct {
-	Id           *int      `json:"id"`
+	Id           *int64    `json:"id"`
 	Date         BasicTime `json:"date"`
 	ShippingCode *string   `json:"shipping_code"`
 	Units        int       `json:"units"`
